% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_algorithm.R
\name{approximate_algorithm}
\alias{approximate_algorithm}
\title{Approximate algorithm}
\usage{
approximate_algorithm(
  W,
  z,
  iteration = 1000,
  a = 1/5,
  b = 10,
  s = 1,
  xi = 1,
  sigma = 1,
  w = 1,
  max_Epsilon = 10^(8),
  tolerance = 1e-20,
  step_check = FALSE
)
}
\arguments{
\item{W}{predictor matrix}

\item{z}{response variance}

\item{iteration}{Number of iterations}

\item{a, b}{rejection sampler parameter}

\item{s}{xi's hyperparameter}

\item{xi}{the initial value of Global shrinkage parameter}

\item{w}{Sigma's hyperparameter}

\item{time_check}{Output run time}

\item{iteration_check}{Output iteration}

\item{beta}{the initial value of beta}

\item{Sigma}{the inial value of Sigma}
}
\value{
beta posterior samples
}
\description{
Run approximate algorithm
}
