% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_algorithm.R
\name{exact_algorithm}
\alias{exact_algorithm}
\title{Exact algorithm}
\usage{
exact_algorithm(
  W,
  z,
  iteration = 1000,
  a = 1/5,
  b = 10,
  s = 0.1,
  xi = 1,
  sigma = 1,
  w = 1,
  step_check = FALSE
)
}
\arguments{
\item{W}{predictor matrix}

\item{z}{response variance}

\item{iteration}{Number of iterations}

\item{a, b}{rejection sampler parameter}

\item{s}{xi's hyperparameter}

\item{xi}{the initial value of Global shrinkage parameter}

\item{w}{Sigma's hyperparameter}

\item{fast_algorithm}{When set to a true value, a calculation is performed to reduce time complexity.}

\item{time_check}{Output run time}

\item{iteration_check}{Output iteration}

\item{beta}{the initial value of beta}

\item{Sigma}{the inial value of Sigma}
}
\value{
beta posterior samples
}
\description{
Run exact algorithm
}
