% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_data.R
\name{make_response}
\alias{make_response}
\title{make_response}
\usage{
make_response(
  W,
  non_zero = 2,
  SD = 5,
  Heteroskedasticity = FALSE,
  fixed_coefficients = NULL
)
}
\arguments{
\item{W}{W matrix}

\item{non_zero}{Number of non_zero signals}
}
\description{
A response variable of the simulation data is generated,
and a value of non-zero coefficients is 10.
}
