% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_data.R
\name{make_simulation_data}
\alias{make_simulation_data}
\title{make_response}
\usage{
make_simulation_data(
  N = 500,
  p = 5000,
  Johndraw = FALSE,
  W_mean = 0,
  W_sd = 1,
  W_scaling = NULL,
  non_zero = 30,
  error_sd = 1,
  fixed_coefficients = NULL
)
}
\arguments{
\item{non_zero}{Number of non_zero signals}

\item{W}{W matrix}
}
\description{
A response variable of the simulation data is generated,
and a value of non-zero coefficients is 10.
}
