% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_data.R
\name{make_W}
\alias{make_W}
\title{make_W}
\usage{
make_W(
  N = 100,
  p = 10,
  cov_matrix = NULL,
  level = vector(mode = "numeric", length = 0)
)
}
\arguments{
\item{N}{Number of rows}

\item{p}{Number of cols}
}
\description{
make simulation data for Scalable Approximate MCMC Algorithms for the Horseshoe Prior
}
