% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_algorithm_integration.R
\name{modified_approximate_algorithm}
\alias{modified_approximate_algorithm}
\title{modified approximate algorithm}
\usage{
modified_approximate_algorithm(
  W,
  z,
  iteration = 5000,
  independence_sigma = FALSE,
  sigma = 1,
  w = 0,
  a = 1/5,
  b = 10,
  fixed_xi = FALSE,
  xi = 1,
  s = 0.01,
  t = 50,
  alpha0 = -0.5,
  alpha1 = -7 * 10^(-4),
  step_check = FALSE
)
}
\arguments{
\item{W}{Design matrix}

\item{z}{Response variable}

\item{iteration}{Number of iterations}

\item{independence_sigma}{Determine whether to remove sigma in the beta prior distribution}

\item{sigma}{error term's variance}

\item{w}{sigma prior's hyperparameter}

\item{a, b}{rejection sampler parameters}

\item{fixed_xi}{Determine whether to fix global shrinkage parameter}

\item{xi}{Global shrinkage parameter}

\item{s}{Variance of log xi when sampling xi}

\item{t}{Interval to run the adaptive probability}

\item{alpha0}{Parameters of the adaptive probability}

\item{alpha1}{Parameters of the adaptive probability}

\item{step_check}{Time measurement at each step of the algorithm}
}
\value{
beta posterior samples
}
\description{
Run modified approximate algorithm
}
