% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rejection_sampler}
\alias{rejection_sampler}
\title{rejection sampler to update local shrinkage parameters lambda.}
\usage{
rejection_sampler(eps, a, b)
}
\description{
rejection sampler to update local shrinkage parameters lambda.
}
\keyword{internal}
