% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Mhorseshoe.R
\docType{package}
\name{Mhorseshoe}
\alias{Mhorseshoe-package}
\alias{Mhorseshoe}
\title{Mhorseshoe: Approximate Algorithm for Horseshoe Prior}
\description{
Mhorseshoe is a package for a high-dimensional Bayesian linear
modeling algorithm using a horseshoe prior. This package provides two
different algorithm functions : \code{\link{exact_horseshoe}},
\code{\link{approx_horseshoe}}. approx_horseshoe is version that can lower
the computational cost than the existing horseshoe estimator through the
approximate MCMC algorithm in the case of \eqn{p >> N} for \eqn{p}
predictors and \eqn{N} observations. You can see examples of the use of the
two algorithms through the vignette, \code{browseVignettes("Mhorseshoe")}.
}
\author{
\strong{Maintainer}: Mingi Kang \email{leehuimin115@g.skku.edu}

Authors:
\itemize{
  \item Kyoungjae Lee \email{leekjstat@gmail.com}
}

}
\keyword{internal}
