% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_algorithm.R
\name{modified_approximate_algorithm}
\alias{modified_approximate_algorithm}
\title{new approximate algorithm_c}
\usage{
modified_approximate_algorithm(
  W,
  z,
  iteration = 5000,
  a = 1/5,
  b = 10,
  w = 1,
  t = 100,
  alpha0 = -0.5,
  alpha1 = -7 * 10^(-4),
  step_check = FALSE
)
}
\arguments{
\item{W}{predictor matrix}

\item{z}{response variance}

\item{iteration}{Number of iterations}

\item{a, b}{rejection sampler parameters}

\item{w}{Sigma's hyperparameter}

\item{alpha0}{Parameters of adaptive probability}

\item{alpha1}{Parameters of adaptive probability}

\item{threshold}{Parameter that determines the active set}

\item{iteration_check}{Output iteration}
}
\value{
beta posterior samples
}
\description{
Run modified approximate algorithm with dependence prior
}
